requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
tensorflow
numpy==1.26.4
Pillow==10.3.0
prometheus-client==0.20.0
python-multipart